requirements.txt
paho-mqtt
pillow